wapp_sentinel/v2/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
httpx>=0.18.2
pydantic>=1.8.2
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
alembic>=1.7.0
pika>=1.2.0
openai>=1.0.0